+++ lib/src/color_logger.dart
@@ -175,6 +175,19 @@ class LoggerHelperFormatter {
           record.stackTrace!,
           methodCount ?? ColorLogger.defaultMethodCounts[record.level] ?? 3,
         );
+        final count =
+            methodCount ?? ColorLogger.defaultMethodCounts[record.level] ?? 3;
+
+        final currentStack = formatStackTrace(StackTrace.current, count);
+        final recordStack = record.stackTrace != null
+            ? formatStackTrace(record.stackTrace!, count)
+            : null;
+
+        stackTraceStr = '''
+$currentStack\n[Current Stack]
+
+${recordStack != null ? '$recordStack\n[Error Stack]' : ''}
+''';
       } else {
         stackTraceStr = formatStackTrace(
           StackTrace.current,